# perf: perform reversal of roots of unity in-place